Fixed everything not working due to the EventHandler, updated some classes.
I fixed the engine crashing because of EventHandler. Apparently Jonathan thought it was a bright idea to assign a new value to a non-const reference when it should have been compared. Oh, well.

I also update some functions in EventHandler and the AudioHandler to use more consts and references, to hopefully improve performance.