Updated Entity to have its own image managing system (not finished) and added the Ship class.
Entity was void. But then there was light, in the form of good constructors.

I added the Ship class, which will ultimately serve as the base for Enemy and Player, but in the meantime is used to sub for Player. It made Window::exec() lot a lot nicer, at least.

I also added some very kawaii colours.