Added a Movable class for entities that move according to angle and speed. Also cleaned up a bit in the code.
I added the Movable class, which has angle and speed and updates according to them at every frame. It is currently the basis of particle and Ship will follow soon.

I also changed Particle's code a bit (hopefully for the better), and added a const header which currently only contains PI.